ai-context-system/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
chromadb==0.4.18
neo4j==5.15.0
minio==7.2.0
pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
pypandoc==1.12
spacy==3.7.2
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2
langchain==0.0.350
langchain-community==0.0.1
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2
networkx==3.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0